Changed the logic for running clients
The renderer/restart script is not used anymore (it was a pain because it included redundant information with the config file + guessing the platform).
Instead, the server can now directly launch the clients thanks to variable substitution in the start/stop/restart properties of the config file + the new domain and env properties.
The downside is that it is now harder to run the clients by hand, because each instance must be run separately, with the proper arguments.

A new script, tools/wildos, makes it easier to start the server (wildos) and stop the clients (wildos stop).
Also, tools/nw has been updated to better match the syntax of the Linux version.
The tools directory should be in the PATH to simplify life.